Move decode/errors to just errors
