+++ packages/codec/lib/abi/decode/abi.ts
@@ -10,7 +10,7 @@ import * as Pointer from "@truffle/codec/pointer";
 import { DecoderRequest, DecoderOptions } from "@truffle/codec/types";
 import * as Evm from "@truffle/codec/evm";
 import { abiSizeInfo } from "@truffle/codec/abi/allocate";
-import { DecodingError, StopDecodingError } from "@truffle/codec/decode/errors";
+import { DecodingError, StopDecodingError } from "@truffle/codec/errors";
 
 type AbiLocation = "calldata" | "eventdata"; //leaving out "abi" as it shouldn't occur here
