+++ packages/codec/lib/errors.ts
